requirements.txt
# PyTorch and related
torch>=2.0.0
numpy>=1.24.0

# Tokenization
sentencepiece>=0.1.99

# Configuration
pyyaml>=6.0

# Progress bars and utilities
tqdm>=4.65.0

# Optional: Logging and visualization
# tensorboard>=2.13.0
# wandb>=0.15.0
